Add files via upload
The frontend of FlutterPromo comprises a user-friendly interface built using the Flutter framework. It offers seamless navigation and intuitive controls for users to interact with the application. Through the frontend, users can effortlessly register for accounts and redeem promo codes during the registration process. The interface provides real-time feedback on code validation and application, ensuring a smooth user experience.